builder: rewrite code to checkout refs
git-checkout $CVSTAG can fail due to other reasons then non existing tag,
i.e. when the merge is in progress. In previous version in such a remote
tracking branch was tried to be check out, which resulted in error. It
prevented user from checking if his merge commits builds correctly
before commiting it.